Reorganize imports to improve performance
