+++ poetry/core/json/__init__.py
@@ -4,8 +4,6 @@ import os
 from io import open
 from typing import List
 
-from jsonschema import Draft7Validator
-
 
 SCHEMA_DIR = os.path.join(os.path.dirname(__file__), "schemas")
 
@@ -23,6 +21,8 @@ def validate_object(obj: dict, schema_name: str) -> List[str]:
 
     with open(schema, encoding="utf-8") as f:
         schema = json.loads(f.read())
+
+    from jsonschema import Draft7Validator
 
     validator = Draft7Validator(schema)
     validation_errors = sorted(validator.iter_errors(obj), key=lambda e: e.path)
