+++ poetry/core/masonry/__init__.py
@@ -6,5 +6,3 @@ A lot of the code used here has been taken from
 `flit <https://github.com/takluyver/flit>`__ and adapted
 to work with the poetry codebase, so kudos to them for showing the way.
 """
-
-from .builder import Builder
