+++ poetry/core/masonry/builder.py
@@ -3,28 +3,28 @@ from typing import TYPE_CHECKING
 from typing import Optional
 from typing import Union
 
-from .builders.sdist import SdistBuilder
-from .builders.wheel import WheelBuilder
-
 
 if TYPE_CHECKING:
     from poetry.core.poetry import Poetry  # noqa
 
 
 class Builder:
-    _FORMATS = {
-        "sdist": SdistBuilder,
-        "wheel": WheelBuilder,
-    }
-
     def __init__(self, poetry: "Poetry") -> None:
+        from .builders.sdist import SdistBuilder
+        from .builders.wheel import WheelBuilder
+
         self._poetry = poetry
 
+        self._formats = {
+            "sdist": SdistBuilder,
+            "wheel": WheelBuilder,
+        }
+
     def build(self, fmt: str, executable: Optional[Union[str, Path]] = None) -> None:
-        if fmt in self._FORMATS:
-            builders = [self._FORMATS[fmt]]
+        if fmt in self._formats:
+            builders = [self._formats[fmt]]
         elif fmt == "all":
-            builders = self._FORMATS.values()
+            builders = self._formats.values()
         else:
             raise ValueError("Invalid format: {}".format(fmt))
