--- poetry/core/masonry/builders/__init__.py
@@ -1,2 +1,0 @@
-from .sdist import SdistBuilder
-from .wheel import WheelBuilder
