+++ poetry/core/packages/project_package.py
@@ -4,19 +4,13 @@ from typing import Dict
 from typing import Optional
 from typing import Union
 
-from poetry.core.semver import VersionRange
-from poetry.core.semver import parse_constraint
+from poetry.core.semver.helpers import parse_constraint
 from poetry.core.version.markers import parse_marker
 
 
 if TYPE_CHECKING:
-    from . import (
-        DirectoryDependency,
-        FileDependency,
-        URLDependency,
-        VCSDependency,
-        Dependency,
-    )
+    from .types import DependencyTypes
+    from poetry.core.semver.helpers import VersionTypes
 
 from .package import Package
 from .utils.utils import create_nested_marker
@@ -26,7 +20,7 @@ class ProjectPackage(Package):
     def __init__(
         self,
         name: str,
-        version: Union[str, VersionRange],
+        version: Union[str, "VersionTypes"],
         pretty_version: Optional[str] = None,
     ) -> None:
         super(ProjectPackage, self).__init__(name, version, pretty_version)
@@ -47,15 +41,7 @@ class ProjectPackage(Package):
     def is_root(self) -> bool:
         return True
 
-    def to_dependency(
-        self,
-    ) -> Union[
-        "DirectoryDependency",
-        "FileDependency",
-        "URLDependency",
-        "VCSDependency",
-        "Dependency",
-    ]:
+    def to_dependency(self) -> Union["DependencyTypes"]:
         dependency = super(ProjectPackage, self).to_dependency()
 
         dependency.is_root = True
@@ -63,11 +49,13 @@ class ProjectPackage(Package):
         return dependency
 
     @property
-    def python_versions(self) -> Union[str, VersionRange]:
+    def python_versions(self) -> Union[str, "VersionTypes"]:
         return self._python_versions
 
     @python_versions.setter
-    def python_versions(self, value: Union[str, VersionRange]) -> None:
+    def python_versions(self, value: Union[str, "VersionTypes"]) -> None:
+        from poetry.core.semver.version_range import VersionRange
+
         self._python_versions = value
 
         if value == "*" or value == VersionRange():
