+++ poetry/core/packages/vcs_dependency.py
@@ -4,8 +4,6 @@ from typing import List
 from typing import Optional
 from typing import Union
 
-from poetry.core.vcs import git
-
 from .dependency import Dependency
 
 
@@ -101,6 +99,8 @@ class VCSDependency(Dependency):
 
     @property
     def base_pep_508_name(self) -> str:
+        from poetry.core.vcs import git
+
         requirement = self.pretty_name
         parsed_url = git.ParsedUrl.parse(self._source)
