+++ tests/spdx/test_license.py
@@ -1,4 +1,4 @@
-from poetry.core.spdx import license_by_id
+from poetry.core.spdx.helpers import license_by_id
 
 
 def test_classifier_name():
